# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "BacktestParameterInput": {
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "value"
        ],
        "title": "BacktestParameterInput",
        "type": "object"
      },
      "BacktestStatus": {
        "enum": [
          "accepted",
          "rejected"
        ],
        "title": "BacktestStatus",
        "type": "string"
      },
      "BacktestSubmissionResponse": {
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/BacktestStatus"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "status",
          "job_id",
          "error_code",
          "error_message",
          "trace_id"
        ],
        "title": "BacktestSubmissionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "JobLogEntryResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "message",
          "created_at"
        ],
        "title": "JobLogEntryResponse",
        "type": "object"
      },
      "JobLogsResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/JobLogEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "JobLogsResponse",
        "type": "object"
      },
      "JobState": {
        "enum": [
          "queued",
          "preparing_data",
          "starting_engine",
          "running",
          "parsing_result",
          "succeeded",
          "failed",
          "cancelled",
          "timed_out"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobStatusResponse": {
        "properties": {
          "artifact_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Uri"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "max_retries": {
            "title": "Max Retries",
            "type": "integer"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "progress_percent": {
            "title": "Progress Percent",
            "type": "integer"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "state",
          "progress_percent",
          "message",
          "updated_at",
          "retry_count",
          "max_retries",
          "artifact_uri"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "MockBacktestRequest": {
        "properties": {
          "adjustment": {
            "enum": [
              "qfq",
              "hfq",
              "none"
            ],
            "title": "Adjustment",
            "type": "string"
          },
          "base_currency": {
            "const": "CNY",
            "title": "Base Currency",
            "type": "string"
          },
          "benchmark": {
            "minLength": 1,
            "title": "Benchmark",
            "type": "string"
          },
          "commission_model": {
            "const": "mock",
            "title": "Commission Model",
            "type": "string"
          },
          "data_snapshot_id": {
            "minLength": 1,
            "title": "Data Snapshot Id",
            "type": "string"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "initial_capital": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Initial Capital"
          },
          "instruments": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Instruments",
            "type": "array"
          },
          "parameters": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BacktestParameterInput"
            },
            "title": "Parameters",
            "type": "array"
          },
          "resolution": {
            "const": "daily",
            "title": "Resolution",
            "type": "string"
          },
          "slippage_model": {
            "const": "mock",
            "title": "Slippage Model",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "strategy_version_id": {
            "minLength": 1,
            "title": "Strategy Version Id",
            "type": "string"
          }
        },
        "required": [
          "strategy_version_id",
          "instruments",
          "start_date",
          "end_date",
          "resolution",
          "adjustment",
          "initial_capital",
          "base_currency",
          "benchmark",
          "commission_model",
          "slippage_model",
          "data_snapshot_id"
        ],
        "title": "MockBacktestRequest",
        "type": "object"
      },
      "MockBacktestResultResponse": {
        "properties": {
          "artifact_uri": {
            "title": "Artifact Uri",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "max_drawdown_percent": {
            "title": "Max Drawdown Percent",
            "type": "number"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "total_return_percent": {
            "title": "Total Return Percent",
            "type": "number"
          },
          "trade_count": {
            "title": "Trade Count",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "state",
          "artifact_uri",
          "total_return_percent",
          "max_drawdown_percent",
          "trade_count"
        ],
        "title": "MockBacktestResultResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "QuantTrade API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/backtests/mock": {
      "post": {
        "operationId": "submit_mock_backtest_backtests_mock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MockBacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Mock Backtest",
        "tags": [
          "backtests"
        ]
      }
    },
    "/backtests/mock/{job_id}/result": {
      "get": {
        "operationId": "get_mock_backtest_result_backtests_mock__job_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MockBacktestResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mock Backtest Result",
        "tags": [
          "backtests"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "system"
        ]
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_status_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "backtests"
        ]
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "backtests"
        ]
      }
    },
    "/jobs/{job_id}/logs": {
      "get": {
        "operationId": "get_job_logs_jobs__job_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobLogsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Logs",
        "tags": [
          "backtests"
        ]
      }
    }
  }
}